requirements.txt
asgiref==3.7.2
crispy-bootstrap4==2022.1
Django==4.2.2
django-crispy-forms==2.0
django-phonenumber-field==7.1.0
phonenumberslite==8.13.14
Pillow==9.5.0
sqlparse==0.4.4
typing_extensions==4.6.3
